fixed issue with larger files

+++ File Server/fileServer_Client.py
@@ -8,34 +8,39 @@ def Main():
     s = socket.socket()
     s.connect((host,port))
 
-    fileName = None
-    while(fileName != 'q'):
-        fileName = input("Enter q to quit or Enter Filename: ")
-        s.sendall(fileName.encode('utf-8'))
+    while True:
+        fileName = input("[+] Enter q to quit or Enter Filename: ")
+        if(fileName != 'q'):
+            s.sendall(fileName.encode('utf-8'))
 
-        serverResponse = s.recv(1024).decode('utf-8')
-        print(serverResponse)
-            
-        if(serverResponse[:2] == 'OK'):
-            regEx = re.compile(r'[0-9]')
-            size = regEx.findall(serverResponse)
-            fileSize = int(size[0])
-            confirm = input("(y/n) >>> ")
-            s.sendall(confirm[0].encode('utf-8'))
+            serverResponse = s.recv(1024).decode('utf-8')
+            print('[!] '+serverResponse)
+                
+            if(serverResponse[:2] == 'OK'):
+                regEx = re.compile(r'Size: [0-9]*')
+                size = regEx.findall(serverResponse)
+                print(size)
+                fileSize = int(size[0].replace('Size: ',''))
+                confirm = input("(y/n) >>> ")
+                s.sendall(confirm[0].encode('utf-8'))
 
-            file = open("downloaded_"+fileName, 'wb')
-            data = s.recv(1024)
-            totalRecv = len(data)
-            file.write(data)
+                if confirm[0] in {'y','Y'}:
+                    file = open("downloaded_"+fileName, 'wb')
+                    data = s.recv(fileSize)
+                    #totalRecv = len(data)
+                    file.write(data)
 
-            while(totalRecv <= fileSize):
-                data = s.recv(1024)
-                totalRecv += len(data)
-                file.write(data)
+                    '''while(totalRecv <= fileSize):
+                        data = s.recv(1024)
+                        totalRecv += len(data)
+                        file.write(data)'''
 
-            print("Download Complete..")
-            file.close()
+                    print("[+] Download Complete!")
+                    file.close()
 
+        else:
+            s.close()
+            exit(0)
     s.close()
     exit(0)
